requirements.txt
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.7.2
dateparser==1.2.0
requests-cache==1.2.1
spacy==3.7.5

# Optional: if you plan to cache responses or add rich CLI colors later you could pin:
# rich==13.9.2
# typer==0.12.5

# After installing, remember to download the spaCy model (once):
# python -m spacy download en_core_web_sm